fix conversion error when assigning charge to particles

+++ src/QDMEwald/src/dummy.cpp
@@ -52,7 +52,7 @@ int main(int argc, char** argv) {
       xyz.at(3 * i + 1) = (double)(iy)*0.5;
       xyz.at(3 * i + 2) = (double)(iz)*0.5;
 
-      q.at(i) = (double)((ix + iy + iz) % 2) ? 1.0 : -1.0;
+      q.at(i) = (((ix + iy + iz) % 2) ? 1.0 : -1.0);
     }
 
     for (double& i : d) i = 0.0;
